Fix return value/source streaming for BwC
